fix: code

+++ Jenkinsfile
@@ -13,6 +13,8 @@ pipeline {
  
         stage('Install sam-cli') {
           steps {
+            File file ='aws-sam-cli-linux-x86_64.zip'
+            if file.exists() {
               sh '''
                rm aws-sam-cli-linux-x86_64.zip
                rm -rf sam-installation
@@ -21,6 +23,7 @@ pipeline {
                ./sam-installation/install --update
                /usr/local/bin/sam --version
             '''
+            }
           }
         }
